requirements.txt
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11